requirements.txt
# Core dependencies (Windows)
setuptools==81.0.0
numpy==2.3.5
requests>=2.31.0,<2.32.0
tqdm
wget

# Audio processing
ffmpeg-python>=0.2.0
faiss-cpu==1.13.2
librosa==0.11.0
scipy==1.16.3
soundfile==0.12.1
noisereduce
pedalboard
stftpitchshift
soxr

# Machine learning and deep learning (Windows CUDA 12.8)
numba==0.63.1
torch==2.7.1+cu128
torchaudio==2.7.1+cu128
torchvision==0.22.1+cu128
torchcrepe
torchfcpe
einops
transformers==4.44.2

# Visualization and UI
matplotlib==3.10.8
gradio==6.5.1
pywebview

# Miscellaneous utilities
beautifulsoup4
sounddevice
webrtcvad-wheels
